requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Quantum Computing
pennylane>=0.32.0
pennylane-lightning>=0.32.0
qiskit>=0.45.0

# Post-Quantum Cryptography
liboqs-python>=0.8.0
pqcrypto>=0.1.0

# Machine Learning & Deep Learning
torch>=1.12.0
tensorflow>=2.10.0
xgboost>=1.6.0

# Web Framework & Dashboard
streamlit>=1.25.0
dash>=2.14.0
plotly>=5.15.0
flask>=2.3.0

# Network Monitoring
scapy>=2.4.5
psutil>=5.9.0
pyshark>=0.6.0

# Data Processing
pyyaml>=6.0
jsonschema>=4.0.0
python-dateutil>=2.8.0
regex>=2022.10.0

# Logging & Monitoring
structlog>=22.0.0
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Security
cryptography>=41.0.0
paramiko>=3.0.0
bcrypt>=4.0.0

# Utilities
click>=8.0.0
tqdm>=4.64.0
colorama>=0.4.0
python-dotenv>=1.0.0

# Development
black>=23.0.0
flake8>=5.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Performance
numba>=0.57.0
joblib>=1.3.0
multiprocess>=0.70.0
